move keystore to client
